+++ src/main/java/konkuk/thip/feed/application/service/BasicFeedShowAllService.java
@@ -8,7 +8,6 @@ import konkuk.thip.feed.application.port.in.FeedShowAllUseCase;
 import konkuk.thip.feed.application.port.out.FeedQueryPort;
 import konkuk.thip.feed.application.port.out.dto.FeedQueryDto;
 import konkuk.thip.post.application.port.out.PostLikeQueryPort;
-import konkuk.thip.saved.application.port.out.SavedQueryPort;
 import lombok.RequiredArgsConstructor;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty;
 import org.springframework.stereotype.Service;
@@ -33,7 +32,6 @@ public class BasicFeedShowAllService implements FeedShowAllUseCase {
 
     private static final int PAGE_SIZE = 10;
     private final FeedQueryPort feedQueryPort;
-    private final SavedQueryPort savedQueryPort;
     private final PostLikeQueryPort postLikeQueryPort;
     private final FeedQueryMapper feedQueryMapper;
 
@@ -50,7 +48,7 @@ public class BasicFeedShowAllService implements FeedShowAllUseCase {
                 .collect(Collectors.toUnmodifiableSet());
 
         // 3. 유저가 저장한 피드들, 좋아한 피드들 조회
-        Set<Long> savedFeedIdsByUser = savedQueryPort.findSavedFeedIdsByUserIdAndFeedIds(feedIds, userId);
+        Set<Long> savedFeedIdsByUser = feedQueryPort.findSavedFeedIdsByUserIdAndFeedIds(feedIds, userId);
         Set<Long> likedFeedIdsByUser = postLikeQueryPort.findPostIdsLikedByUser(feedIds, userId);
 
         // 4. response 로의 매핑
