+++ src/test/java/konkuk/thip/feed/adapter/in/web/FeedShowSpecificUserApiTest.java
@@ -8,8 +8,8 @@ import konkuk.thip.feed.adapter.out.persistence.repository.Content.ContentJpaRep
 import konkuk.thip.feed.adapter.out.persistence.repository.FeedJpaRepository;
 import konkuk.thip.post.adapter.out.jpa.PostLikeJpaEntity;
 import konkuk.thip.post.adapter.out.persistence.PostLikeJpaRepository;
-import konkuk.thip.saved.adapter.out.jpa.SavedFeedJpaEntity;
-import konkuk.thip.saved.adapter.out.persistence.repository.SavedFeedJpaRepository;
+import konkuk.thip.feed.adapter.out.jpa.SavedFeedJpaEntity;
+import konkuk.thip.feed.adapter.out.persistence.repository.SavedFeedJpaRepository;
 import konkuk.thip.user.adapter.out.jpa.AliasJpaEntity;
 import konkuk.thip.user.adapter.out.jpa.UserJpaEntity;
 import konkuk.thip.user.adapter.out.persistence.repository.UserJpaRepository;
